feat: build and run the uc CLI on Windows
The uc CLI imports the daemon's internal/machine package, which previously
only compiled on Linux and darwin. Add Windows support following the existing
darwin client-only pattern so the CLI builds and runs natively on Windows.

- Add windows stubs for the firewall, docker controller, and wireguard
  network packages, mirroring the darwin stubs.
- Split listenUnixSocket into machine_unix.go (real impl) and
  machine_windows.go (stub), since docker/go-connections sockets.NewUnixSocket
  is Unix-only.
- Extract terminal resize signal handling into client_exec_{unix,windows}.go
  because unix.SIGWINCH does not exist on Windows.
- Disable SSH connection multiplexing (ControlMaster/ControlPath) on Windows
  in the SSH CLI connector. Windows OpenSSH does not support the Unix-domain
  control socket and fails with "getsockname failed: Not a socket".

Caveats:
- The uncloudd daemon is not supported on Windows. The Windows daemon-side
  code is stubs only, present so the CLI packages compile; calling it returns
  "not supported on Windows" errors. Windows is a client-only target, like
  darwin.
- With SSH multiplexing disabled, Windows reuses no connections, so every
  gRPC dial spawns a fresh ssh process. This is slower than on Linux/macOS
  but fully functional.
