+++ internal/machine/docker/client_exec_windows.go
@@ -1,0 +1,7 @@
+package docker
+
+import "os"
+
+// notifyResizeSignal is a no-op on Windows, which does not deliver terminal
+// resize notifications via signals. The initial window size is still sent.
+func notifyResizeSignal(ch chan<- os.Signal) {}
