+++ internal/machine/network/wireguard_windows.go
@@ -1,0 +1,30 @@
+//go:build windows
+
+package network
+
+import (
+	"context"
+	"errors"
+)
+
+type WireGuardNetwork struct{}
+
+func NewWireGuardNetwork() (*WireGuardNetwork, error) {
+	return &WireGuardNetwork{}, nil
+}
+
+func (n *WireGuardNetwork) Configure(config Config) error {
+	return errors.New("not implemented on windows")
+}
+
+func (n *WireGuardNetwork) Run(ctx context.Context) error {
+	return errors.New("not implemented on windows")
+}
+
+func (n *WireGuardNetwork) WatchEndpoints() <-chan EndpointChangeEvent {
+	return nil
+}
+
+func (n *WireGuardNetwork) Cleanup() error {
+	return errors.New("not implemented on windows")
+}
